requirements.txt
fastapi
uvicorn
mangum
motor
python-dotenv
semantic-kernel==1.35.0
semantic-kernel[google]==1.35.0
azure-ai-projects>=1.0.0b11